Add files via upload
Analysis of fraud detection on a insurance data. It is checked whether a customer has raised for a valid or fraud detection using several KPIs(features). Since the target variable is  categorical variable and its a classification problem, so decision tree, random forest, KNN and several similar algorithms are used.